% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_threshold_aggregate.R
\name{aggregate_df}
\alias{aggregate_df}
\title{Aggregate Dataframe}
\usage{
aggregate_df(df, aggregation, date, lat, lon)
}
\arguments{
\item{df}{The binary time-frequency dataframe of acoustic activity (active=1, inactive=0) produced by \code{\link{binarize_df}}}

\item{aggregation}{The time interval by which binarized spectral index vectors are pooled. Options (in minutes) are 1, 5, 10, 20, 30 and 60.
If non-continuous recordings are used, make sure the interval between recordings does not exceed the aggregation interval.}

\item{date}{The first day of the recording period. Used for managing time-objects in R. Format as "YYYY-mm-dd".}

\item{lat}{The latitude of the site at which the sound files were collected.}

\item{lon}{The longitude of the site at which the sound files were collected.}
}
\value{
Returns an aggregated time-frequency dataframe of the proportion of acoustically active recordings
in each bin for a set acoustic index.
}
\description{
Pools binarized spectral indices by time of day using a user-defined aggregation interval.
For each aggregate time period, the activity values (0 or 1) are summed per frequency bin, and divided by
the number of recordings for that time interval to get the proportion of acoustically active recordings in
each time-frequency bin.
}
