% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amazon_soundscape_ACI}
\alias{amazon_soundscape_ACI}
\title{Soundscape data from the Brazilian Amazon - Acoustic Complexity Index}
\format{
A time-frequency data frame with 56 rows and 1603 columns. Rows represent frequency bins and columns show the time
of day when the recording was taken. For each time-frequency bin, the 'Acoustic Complexity Index' value is given.
}
\source{
Amazon Biodiversity & Carbon Expeditions (unpublished)
}
\usage{
data(amazon_soundscape_ACI)
}
\description{
Soundscape data from the Brazilian Amazon, collected at the Cristalino Ecolodge in November 2019.
The data was collected using an AudioMoth recorder, sampling the soundscape for 1-min every 10-min at a
384,000 Hz sampling rate and medium gain for approximately 12 days. This resulted in 1603 sound files.
This raw sound data was processed using the \code{soundscapeR} package. For each file, a set of spectral
acoustic indices was computed using the \code{\link{index_calc}} function. Next, all spectral index '.csv' files
for the Acoustic Complexity Index ('ACI') were merged into a single time-frequency dataframe of index values. Finally,
dataframe was subsetted to include frequencies between 0-21,000 Hz.
}
\author{
Thomas Luypaert \email{thomas.luypaert@nmbu.no}
}
\keyword{data}
\keyword{datasets}
