% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_binarize_df.R
\name{binarize_df}
\alias{binarize_df}
\title{Binarize Dataframe}
\usage{
binarize_df(merged_soundscape, method, value = NULL)
}
\arguments{
\item{merged_soundscape}{The merged soundscape object produced by
\code{\link{merge_csv}}.}

\item{method}{The algorithm used to determine the threshold.
Options are "IJDefault","Huang", "Huang2", "Intermodes",
"IsoData", "Li", "MaxEntropy", "Mean", "MinErrorI", "Minimum",
"Moments", "Otsu", "Percentile", "RenyiEntropy", "Shanbhag",
"Triangle", "Yen", and "Mode". To specify a custom threshold,
use method="Custom" in combination with the value argument.
Consult \url{http://imagej.net/Auto_Threshold} for more
information on algorithm methodologies.}

\item{value}{Optional argument used to set a custom threshold
value for binarization - used in combination with method="Custom".}
}
\value{
Returns a binary time-frequency dataframe of acoustic
 activity (active=1, inactive=0) for a set acoustic index.
}
\description{
Separates acoustically active time-frequency bins
 (active=1) from background values (inactive=0) through the
 application of a binarization algorithm. Several binarization
 algorithms are available, either based on image thresholding
 using \code{\link[autothresholdr]{auto_thresh}}, or subtraction
 of the mode.
}
