% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_coords.R
\name{circ_scale}
\alias{circ_scale}
\title{Calculated scaling function for a custom range of response values}
\usage{
circ_scale(range)
}
\arguments{
\item{range}{A double with the min and max possible values of the response, i.e. \code{c(1, 5)}}
}
\value{
A float
}
\description{
Calculated scaling function for a custom range of response values
}
