% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_function_names.R
\name{defunct}
\alias{defunct}
\title{Defunct function}
\usage{
defunct(msg = "This function is deprecated")
}
\arguments{
\item{msg}{Message that is returned when a deprecated function is used.}
}
\description{
Throws an error message when depreciated function names are used
}
