% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_binarize_df.R
\name{get_mode}
\alias{get_mode}
\title{Calculate the mode}
\usage{
get_mode(df)
}
\arguments{
\item{df}{The merged_soundscape data frame produced by
\code{\link{merge_csv}}.}
}
\value{
Returns the acoustic index value which appears most often
}
\description{
Calculate the mode
}
\details{
Function for internal use by \code{\link{threshold_df}}
}
