% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_ss_index_calc.R
\name{get_wav_duration}
\alias{get_wav_duration}
\title{Check the duration of a sound file}
\usage{
get_wav_duration(file_path)
}
\arguments{
\item{file_path}{A full length path to the location of a sound file.}
}
\value{
The duration of a .wav sound file in seconds.
}
\description{
This function takes the full length path to a sound file and returns its duration in seconds.
}
\keyword{internal}
