% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{heatmapper}
\alias{heatmapper}
\title{Flexible Soundscape Heatmaps}
\usage{
heatmapper(
  df,
  type = "regular",
  annotate = TRUE,
  timeinterval = "1 hour",
  mintime = "default",
  maxtime = "default",
  freqinterval = 1000,
  minfreq = 0,
  maxfreq = "default",
  nbins = 10,
  date,
  lat,
  lon,
  twilight = "sunlight",
  labelsize_time = 4,
  labelsize_frequency = 4,
  labelsize_polar = 3,
  palette = "D",
  direction = 1,
  marginplot = FALSE,
  interactive = FALSE,
  save = FALSE,
  dir = "default",
  filename = "file",
  device = "png"
)
}
\arguments{
\item{df}{The aggregated time-frequency dataframe produced by \code{\link{aggregate_df}}.}

\item{type}{One of either "regular" or "polar". If set to "regular", produces a regular rectangular heatmap.
If set to "polar", produces a polar heatmap suitable for exploring diurnal patterns.}

\item{annotate}{One of either TRUE or FALSE. If set to TRUE, annotates the heatmap with sunrise and sunset times,
and highlights the border between the audible and ultrasonic spectrum for human hearing.}

\item{timeinterval}{A time interval for the x-axis. Options can be found in the \code{\link[scales]{date_breaks}} documentation.}

\item{mintime}{The lower time limit for the x-axis, formatted as "HH:MM:SS". Defaults to the earliest
time for which data exists in the dataframe.}

\item{maxtime}{The upper time limit for the x-axis, formatted as "HH:MM:SS".Defaults to the latest
time for which data exists in the dataframe.}

\item{freqinterval}{The frequency interval for the y-axis, expressed as a numeric value.}

\item{minfreq}{The lower frequency limit for the y-axis as a numeric value. Defaults to zero.}

\item{maxfreq}{The upper frequency limit for the y-axis as a numeric value. Defaults to the maximum frequency of the dataframe.}

\item{nbins}{If \code{marginplot=TRUE}, determines the number of the frequency-bins by which to divide the frequency range to compute the
soundscape richness, expressed as a numeric constant.}

\item{date}{The first day of the recording period. Used for managing time-objects in R.  Formatted as "YYYY-mm-dd".}

\item{lat}{The latitude of the site at which the sound files were collected.}

\item{lon}{The longitude of the site at which the sound files were collected.}

\item{twilight}{A character string of the twilight method to be used for sunrise and sunset annotation.
Options can be found in the \code{\link[photobiology]{day_night}} documentation.}

\item{labelsize_time}{If annotate=TRUE, can be used to alter the size of temporal spectrum annotation.}

\item{labelsize_frequency}{If annotate=TRUE, can be used to alter the size of frequency spectrum annotation.}

\item{labelsize_polar}{If type="polar", can be used to alter the size of the frequency axis labels.}

\item{palette}{A character string indicating the colormap option to use.
Four options are available: "magma" (or "A"), "inferno" (or "B"), "plasma" (or "C"),
"viridis" (or "D", the default option) and "cividis" (or "E"). Consult \url{https://www.rdocumentation.org/packages/viridisLite/versions/0.3.0/topics/viridis} for options.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, the regular order is followed.
If -1, the order of colors is reversed.}

\item{marginplot}{One of either TRUE or FALSE. If set to TRUE, adds marginal plots to the x- and y-axes.
For the x-axis, the marginal plot displays the smoothed richness of acoustically active frequency bins for each time of day.
For the y-axis, the marginal plot displays the smoothed richness of acoustically active time-bins for each frequency band.
Note that marginal plots are not available for type="polar".}

\item{interactive}{One of either TRUE or FALSE. If set to TRUE, an interactive plot is produced using \code{\link[plotly]{ggplotly}}.
Note that interactive plots are not available for marginplot=TRUE.}

\item{save}{One of either TRUE or FALSE. If set to TRUE, saves the plot using \code{\link[ggplot2]{ggsave}}, and the 'dir', 'filename' and 'device'
arguments.}

\item{dir}{Path of the directory to save plot to: path and filename are combined to create the fully qualified file name.
Defaults to the working directory. For more information consult \code{\link[ggplot2]{ggsave}}.}

\item{filename}{The file name without the extention. For more information consult \code{\link[ggplot2]{ggsave}}.}

\item{device}{Device to use. Can either be a device function (e.g. png()), or one of
"eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).
Defaults to "png". For more information consult \code{\link[ggplot2]{ggsave}}.}
}
\value{
Returns a ggplot heatmap object and if save=TRUE, saves the plot in a directory of choice using a specified device and filename.
}
\description{
Creates a soundscape heatmap to visualize the use of acoustic
space in the time-frequency domain for an set acoustic index. The function is highly
flexible, allowing the user maximal control over every aspect of the heatmap. Please
consult the arguments section to find out more about visualization options.
}
