% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_calculate_indices.R
\name{index_calc}
\alias{index_calc}
\title{Compute Acoustic Indices}
\usage{
index_calc(fileloc, progloc, samplerate = 41000, window = 256)
}
\arguments{
\item{fileloc}{The full-length path to the sound file for which to
compute indices.}

\item{progloc}{The full-length path to the location of
'AnalysisPrograms.exe'.}

\item{samplerate}{The number of times the sound was sampled each second.
This is a fixed parameter determined by your recording setup, although
downsampling to a lower sampling rate is possible.}

\item{window}{A variable of the Short-time Fourier Transformation,
expressed as the number of samples. The window size of choice depends on
the fundamental frequency, intensity and change of the signal of interest,
and influences the temporal and frequency resolution of the analysis.
The window size is generally a power of 2.}
}
\value{
Creates a directory with '.csv' files for the aforementioned
set of spectral and summary indices. The directory is created in the same
place as the original sound file.
}
\description{
Calls on the external software 'AnalysisPrograms.exe'
to compute a set of acoustic indices using user-specified
configuration parameters.
}
\details{
The default duration of sound files for index computation is 60 seconds.
If the file length exceeds 60 seconds, the file is automatically cut into
60 second segments for further analysis.

\bold{The following spectral indices will be computed:}

\emph{Background Noise (BGN)}:
The mode of the distribution of decibel values in each
frequency bin, representing the “background” intensity value.
This index captures the acoustic energy which persists throughout the
duration of the sound file, regardless of its origin
(biophonic, geophonic or anthrophonic).

\emph{Power minus noise (PMN)}:
The difference between the maximum decibel value in
each frequency bin and the corresponding BGN decibel value.

\emph{Acoustic cover (CVR)}:
The fraction of active elements in each noise-reduced
frequency bin where the amplitude exceeds a 3-dB threshold.

\emph{Number of events (EVN)}:
The number of times the decibel value in a noise-reduced
frequency bin crosses the 3-dB threshold from lower to higher values.

\emph{Temporal entropy (ENT)}:
A measure of acoustic energy concentration in each noise-reduced
frequency bin.

\emph{Acoustic Complexity Index (ACI)}:
A measure quantifying the variability in
intensity values in each noise-reduced frequency bin. It is widely used as a
measure of biophony in recordings, however remains highly sensitive to
non-biological sources of sound.

\emph{Oscillation Index (OSC)}:
...

\emph{Spectral Peak Tracks (SPT)}:
A measure of the presence of spectral peak
tracks in a noise-reduced frequency bin.

\emph{Ridge indices (RHZ, RVT, RPS, RNS)}:
A set of indices based on the presence of formants
in the harmonic structure of many animal vocalizations, calculated in the
four directions of the ridge slope (horizontal (RHZ), vertical (RVT),
upward slope (RPS), downward slope (RNS)). Formants in the mid-band are
typically due to birdsong, whereas vertical formants are typical for
non-biological sounds such as rain drops and electrical clicks.

For index computation this function calls on the external software
'AnalysisPrograms.exe' by the QUT Ecoacoustics Research Group
(\url{https://github.com/QutEcoacoustics/audio-analysis}).Make sure the
software is installed prior to commencing the analysis. Additionally, the
software requires sound files to be named according to certain standards
(\url{https://ap.qut.ecoacoustics.info/basics/dates.html}).

\strong{Software citation}:
Michael Towsey, Anthony Truskinger, Mark Cottman-Fields, & Paul Roe.
(2018, March 5). Ecoacoustics Audio Analysis Software v18.03.0.41
(Version v18.03.0.41). Zenodo. http://doi.org/10.5281/zenodo.1188744
}
