% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indices.R
\name{index_config}
\alias{index_config}
\title{Set Configuration Parameters for Acoustic Index Computation}
\usage{
index_config(progloc, indexlength = 60, samplerate = 41000, window = 256)
}
\arguments{
\item{progloc}{The full-length path to the location of the 'AnalysisPrograms.exe' software.}

\item{indexlength}{The timespan (in seconds) over which the acoustic indices are computed.
This value cannot exceed 60 seconds, as this is the default file length for index computation.}

\item{samplerate}{The number of times the sound was sampled each second. This is a fixed
parameter determined by your recording setup, although downsampling to a lower sampling
rate is possible.}

\item{window}{A variable of the Short-time Fourier Transformation, expressed as the number
of samples. The window size of choice depends on the fundamental frequency, intensity and change
of the signal of interest, and influences the temporal and frequency resolution of the analysis.
The window size is generally a power of 2.}
}
\value{
Edits and saves the '.yml' configuration file which is used by the AnalysisPrograms software for
index computation.
}
\description{
Alters and saves the configuration file used by external software 'AnalysisPrograms.exe'
for the computation of acoustic indices.
}
\details{
The \code{index_config} function is normally only used internally by the \code{\link{index_calc}} function,
although it is possible for the user to edit the configuration file manually.
}
