% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISD.R
\docType{data}
\name{ISD}
\alias{ISD}
\title{International Soundscape Database}
\format{
\subsection{\code{ISD}}{

A data frame with 1,909 rows and 78 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
\url{https://zenodo.org/record/6331810}
}
\usage{
ISD
}
\description{
A database of soundscape surveys and acoustic analysis results from Europe
Report ...
}
\keyword{datasets}
