% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_coords.R
\name{iso_coords}
\alias{iso_coords}
\title{Calculate the projected ISOPleasant and ISOEventful coordinates.}
\usage{
iso_coords(df, scale_to_one = TRUE, val_range = c(1, 5), projection = TRUE)
}
\arguments{
\item{df}{Dataframe containing ISD formatted data}

\item{scale_to_one}{Should the coordinates be scaled to (-1, +1)}

\item{val_range}{The range of values, by default Likert scaled c(1, 5)}

\item{projection}{Use the trigonometric projection (cos(45)) term for diagonal PAQs}
}
\value{
A list of 2 with ISOPleasant and ISOEventful values
}
\description{
Calculate the projected ISOPleasant and ISOEventful coordinates.
}
\examples{
df <- tibble::tibble(
  pleasant = c(2, 3, 4), chaotic = c(4, 4, 3), vibrant = c(3, 2, 3),
  uneventful = c(2, 1, 2), calm = c(1, 1, 2), annoying = c(2, 4, 4),
  eventful = c(4, 2, 3), monotonous = c(3, 4, 2)
)
iso_coords(df, val_range = c(1, 5))
}
