% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOjointplot.R
\name{ISOjointplot}
\alias{ISOjointplot}
\title{Create a jointplot with 2D density in the center and distributions on the margins.}
\usage{
ISOjointplot(df, x = "ISOPleasant", y = "ISOEventful", group = "LocationID")
}
\arguments{
\item{df}{Dataframe including ISO coordinate data}

\item{x}{Column name for x}

\item{y}{Column name for y}

\item{group}{Column name of the grouping variable}
}
\value{
ggplot
}
\description{
Create a jointplot with 2D density in the center and distributions on the margins.
}
