% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{lengthen}
\alias{lengthen}
\title{Reformats the dataframe for subsequent visualization}
\usage{
lengthen(df, date, lat, lon)
}
\arguments{
\item{df}{The aggregated time-frequency dataframe produced by \code{\link{aggregate_df}}.}

\item{date}{The first day of the recording period. Used for managing time-objects in R. Format as "YYYY-mm-dd".}

\item{lat}{The latitude of the site at which the sound files were collected.}

\item{lon}{The longitude of the site at which the sound files were collected.}
}
\value{
Returns a long dataframe with "frequency", "time" and "value" columns.
}
\description{
Lengthens the dataframe so it can be used for subsequent visualization using ggplot2.
Function intended for internal use by \code{\link{heatmapper}}, however can also be used manually by user.
}
