% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadISD.R
\name{loadISD}
\alias{loadISD}
\title{Load the International Soundscape Database (ISD)}
\usage{
loadISD(version = "latest")
}
\arguments{
\item{version}{A string with the ISD version number}
}
\value{
A tibble
}
\description{
Load the International Soundscape Database (ISD)
}
\examples{
df <- loadISD(version = "0.2.4")
head(df)
}
