% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_threshold_aggregate.R
\name{merge_csv}
\alias{merge_csv}
\title{Import and Merge Acoustic Index '.csv' Files}
\usage{
merge_csv(fileloc, samplerate, window, index, lat, lon)
}
\arguments{
\item{fileloc}{The full-length path to the directory where the output directories of \code{index_calc} are located.}

\item{samplerate}{The sampling rate specified in the \code{index_calc} function.}

\item{window}{The window length specified in the \code{index_calc} function.}

\item{index}{The acoustic index of interest. Options are "BGN", "PMN", "CVR", "EVN",
"ENT", "ACI", "OSC", "SPT", "RHZ", "RVT", "RPS" and "RNS". For a brief description of indices,
consult the \code{\link{index_calc}} documentation}

\item{lat}{The latitude of the site at which the sound files were collected.}

\item{lon}{The longitude of the site at which the sound files were collected.}
}
\value{
Returns a time-by-frequency dataframe of acoustic index values for all files in the recording period.
}
\description{
For an acoustic index of choice, imports all spectral index '.csv' files
produced by the \code{\link{index_calc}} function, and merges them into a time-by-frequency dataframe.
}
