% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_threshold_aggregate.R
\name{normalize_index}
\alias{normalize_index}
\title{Normalize index values using min-max normalization}
\usage{
normalize_index(df)
}
\arguments{
\item{df}{The time-frequency dataframe of index values produced by \code{merge_csv}}
}
\value{
Returns a time-by-frequency dataframe of nornalized acoustic index values.
}
\description{
Normalizes the index values to scale between 0-1 using min-max normalization.
}
