% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_osu_matrix.R
\name{osu_matrix}
\alias{osu_matrix}
\title{Calculate and OSU-by-incidence matrix for the soundscape object}
\usage{
osu_matrix(split_soundscape)
}
\arguments{
\item{split_soundscape}{The soundscape object obtained after using the 'split_df' function.}
}
\value{
Returns a soundscape object containing a new OSU-by-incidence matrix
}
\description{
Calculate and OSU-by-incidence matrix for the soundscape object
}
