% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{quick_heatmap}
\alias{quick_heatmap}
\title{Quick heatmap}
\usage{
quick_heatmap(df)
}
\arguments{
\item{df}{A time-frequency dataframe produced by \code{\link{merge_csv}}, \code{\link{binarize_df}} or \code{\link{aggregate_df}}.}
}
\value{
Returns a heatmap plot
}
\description{
A quick-and-dirty interactive heatmap for rapid data checking, using the \code{\link[plsgenomics]{matrix.heatmap}} function.
Used for quick visual inspection of outputs for the \code{\link{merge_csv}} and \code{\link{binarize_df}} functions.
Useful for checking which binarization algorithm works best, and fine-tuning threshold values.
}
