% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizualize_diversity_metrics.R
\name{richness_by_time}
\alias{richness_by_time}
\title{Visualize Soundscape Richness by Time of Day}
\usage{
richness_by_time(
  df,
  graphtype = "total",
  date,
  lat,
  lon,
  minfreq = "default",
  maxfreq = "default",
  nbins = 10,
  timeinterval = "1 hour",
  smooth = TRUE,
  movavg = 6,
  interactive = TRUE,
  save = FALSE,
  dir = "default",
  filename = "file",
  device = "png"
)
}
\arguments{
\item{df}{The aggregated time-frequency dataframe produced by \code{\link{aggregate_df}}.}

\item{graphtype}{The type of plot which is produced. \cr
\cr
\strong{Options are}:\cr
\cr
\bold{\emph{'total'}}:
\cr
An area chart showing the soundscape richness by time-of-day for the entire frequency range.
\cr
\bold{\emph{'frequency'}}:
\cr
A stacked area chart showing the relative contribution of frequency bins with user-defined width to the total soundscape richness by time-of-day.
\cr
\bold{\emph{'normfreq'}}:
\cr
A percentage stacked area chart showing the normalized relative contribution of frequency bins with user-defined width to the soundscape richness by time-of-day.
\cr
\bold{\emph{'linefreq'}}:
\cr
A line chart showing the relative contribution of frequency bins with user-defined width to the soundscape richness by time-of-day.}

\item{date}{The first day of the recording period. Used for managing time-objects in R. \cr
Format as "YYYY-mm-dd".}

\item{lat}{The latitude of the site at which the sound files were collected.}

\item{lon}{The longitude of the site at which the sound files were collected.}

\item{minfreq}{The lower frequency limit for which to visualize the soundscape richness, expressed as a numeric value.
Defaults to the lowest frequency for which data exists in the dataframe.}

\item{maxfreq}{The upper frequency limit for which to visualize the soundscape richness, expressed as a numeric value.
Defaults to the highest frequency for which data exists in the dataframe.}

\item{nbins}{If graphtype='frequency'/'normfreq'/'linefreq', determines the number of the frequency-bins by which to divide the frequency range to compute the
relative contribution of each bin to the total richness.}

\item{timeinterval}{A time interval for the x-axis. Options can be found in the \code{\link[scales]{date_breaks}} documentation.}

\item{smooth}{One of either TRUE or FALSE. If set to TRUE, applies a moving average filter for smoothing the richness by time-of-day.}

\item{movavg}{If smooth=TRUE, determines the width of the moving average filter. Consult \code{\link[pracma]{movavg}} for more information.}

\item{interactive}{One of either TRUE or FALSE. If set to TRUE, an interactive plot is produced using \code{\link[plotly]{ggplotly}}.}

\item{save}{One of either TRUE or FALSE. If set to TRUE, saves the plot using \code{\link[ggplot2]{ggsave}}, and the 'dir', 'filename' and 'device'
arguments.}

\item{dir}{Path of the directory to save plot to: path and filename are combined to create the fully qualified file name.
Defaults to the working directory. For more information consult \code{\link[ggplot2]{ggsave}}.}

\item{filename}{The file name without the extention. For more information consult \code{\link[ggplot2]{ggsave}}.}

\item{device}{Device to use. Can either be a device function (e.g. png()), or one of
"eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).
Defaults to "png". For more information consult \code{\link[ggplot2]{ggsave}}.}
}
\value{
Returns a ggplot object and if save=TRUE, saves the plot in a directory of choice using a specified device and filename.
}
\description{
Produces plots showing the variation in soundscape richness by time-of-day. Richness
can be shown for the full frequency range, or the relative contribution of frequency-bins with
user-specified width.
}
