% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_function_names.R
\name{sounddiv_by_time}
\alias{sounddiv_by_time}
\title{sounddiv_by_time function}
\usage{
sounddiv_by_time(...)
}
\arguments{
\item{...}{no parameters}
}
\description{
Depreciated - Use 'ss_diversity_plot' instead
}
