% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundscape_diversity_calculations.R
\name{sounddiv_internal}
\alias{sounddiv_internal}
\title{Estimate Soundscape Diversity using Hill Numbers}
\usage{
sounddiv_internal(
  df,
  qvalue,
  type = "total",
  date,
  lat,
  lon,
  minfreq = "default",
  maxfreq = "default",
  mintime = "default",
  maxtime = "default",
  twilight = "sunlight",
  dawnstart = 0,
  dawnend = 5400,
  duskstart = 5400,
  duskend = 0,
  freqseq = FALSE,
  nbins = 10,
  output = "percentage"
)
}
\arguments{
\item{df}{The aggregated time-frequency dataframe produced by \code{\link{aggregate_df}}.}

\item{qvalue}{A parameter which modulates the sensitivity of diversity values to the relative abundance of Operational Sound Units (OSUs). A positive integer or decimal number (>=0), most commonly between 0-3. A value of 0 corresponds to the richness, a value of 1 is the equivalent number of effective OSUs for the Shannon index, a value of 2 is the equivalent number of effective species for the Simpson index.}

\item{type}{The scale for which the soundscape diversity is computed. Options are 'total', 'day',
'night', 'dawn', 'dusk' and 'tod' (time of day - for each unique time in the day).}

\item{date}{The first day of the recording period. Used for managing time-objects in R.
Format as "YYYY-mm-dd".}

\item{lat}{The latitude of the site at which the sound files were collected.}

\item{lon}{The longitude of the site at which the sound files were collected.}

\item{minfreq}{The lower frequency limit for which to compute the soundscape diversity, expressed as a numeric value.
Defaults to the lowest frequency for which data exists in the dataframe.}

\item{maxfreq}{The upper frequency limit for which to compute the soundscape diversity, expressed as a numeric value.
Defaults to the highest frequency for which data exists in the dataframe.}

\item{mintime}{The lower time limit for which to compute the soundscape diversity, formatted as "HH:MM:SS". Defaults to the
earliest time for which data exists in the dataframe.}

\item{maxtime}{The upper time limit for which to compute the soundscape diversity, formatted as "HH:MM:SS". Defaults to the
latest time for which data exists in the dataframe.}

\item{twilight}{A character string of the twilight method to be used for sunrise and sunset as the boundary between day and night.
Options can be found in the \code{\link[photobiology]{day_night}} documentation.}

\item{dawnstart}{If type='dawn', used to determine the start of dawn. By default, dawn starts at sunrise. Expressed as the
time in seconds before sunrise.}

\item{dawnend}{If type='dawn', used to determine the end of dawn. By default, dawn ends 1.5 hours after sunrise. Expressed as the
time in seconds after sunrise.}

\item{duskstart}{If type='dusk', used to determine the start of dusk. By default, dusk starts 1.5 hours before sunset. Expressed as the
time in seconds before sunset.}

\item{duskend}{If type='dusk', used to determine the end of dusk. By default, dusk ends at sunset. Expressed as the
time in seconds after sunset.}

\item{freqseq}{One of either TRUE or FALSE. If set to FALSE, will compute the diversity for the entire frequency range of the soundscape.
If set to TRUE, will compute the diversity per frequency-bin of user-defined width.}

\item{nbins}{If freqseq is set to TRUE, determines the number of the frequency-bins by which to divide the frequency range to compute the
soundscape diversity, expressed as a numeric constant.}

\item{output}{The format in which the soundscape diversity is expressed. Options are "percentage"
(the fraction between the observed soundscape diversity and the maximum possible soundscape diversity), or "raw"
(the number of acoustically active OSUs in the soundscape). Defaults to "percentage".}
}
\value{
Depending on the chosen parameters, returns the soundscape diversity either a numeric value, a vector of values or a list of vectors of values.
}
\description{
Same as soundscape_Hill, but modified for simpler internal use by sounddiv_by_time
}
