% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_partitioning.R
\name{sounddiv_pairdis}
\alias{sounddiv_pairdis}
\title{Pairwise Beta diversity and dissimilarity values between subsystems in the system}
\usage{
sounddiv_pairdis(
  df_list,
  qvalue,
  hier_table = "default",
  level = "default",
  type = "total",
  minfreq = "default",
  maxfreq = "default",
  mintime = "default",
  maxtime = "default",
  date,
  lat,
  lon,
  twilight = "sunlight",
  dawnstart = 0,
  dawnend = 5400,
  duskstart = 5400,
  duskend = 0,
  output = "percentage"
)
}
\arguments{
\item{df_list}{A list of dataframes of equal dimensions, each dataframe being produced by \code{\link{aggregate_df}}.}

\item{qvalue}{A positive integer or decimal number (>=0), most commonly between 0-3. This parameter modulates the sensitivity of diversity values to the relative abundance of Operational Sound Units (OSUs). A value of 0 corresponds to the richness, a value of 1 is the equivalent number of effective OSUs for the Shannon index, a value of 2 is the equivalent number of effective OSUs for the Simpson index.}

\item{hier_table}{A matrix indicating the relationship between the soundscapes in the df_list. The first column lists the names of all the soundscapes in the df_list, other columns can be used to group soundscapes into higher hierarchical levels. If no hierarchy table is supplied, the function defaults to a 2-level diversity partitioning.}

\item{level}{A positive integer. If a multi-level hierarchical structure exists in the data, as supplied in the hierarchy table, this argument indicates the hierarchical level for which to compute the pairwise values. level=1 computes the pairwise values at the lowest hierarchical level (sub-systems or individual soundscapes), whereas level=2,...,X refers to higher levels of grouping, with the level value corresponding to the column number of the grouping variable in the hierarchy table.}

\item{type}{The scale for which the soundscape diversity is computed. Options are 'total', 'day', night', 'dawn', 'dusk' and 'tod' (time of day - for each unique time in the day).}

\item{minfreq}{A numeric value indicating the lower frequency limit for which to compute the soundscape diversity. If set to default, uses the lowest available frequency in the dataframe.}

\item{maxfreq}{A numeric value indicating the upper frequency limit for which to compute the soundscape diversity. If set to default, uses the highest available frequency in the dataframe.}

\item{mintime}{The lower time limit for which to compute the soundscape diversity, formatted as "HH:MM:SS". If set to default, uses the earliest time for which data exists in the dataframe.}

\item{maxtime}{The upper time limit for which to compute the soundscape diversity, formatted as "HH:MM:SS". If set to default, uses the latest time for which data exists in the dataframe.}

\item{date}{The first day of the recording period. Used for managing time-objects in R.
Format as "YYYY-mm-dd".}

\item{lat}{The latitude of the site at which the sound files were collected, expressed in decimal degrees.}

\item{lon}{The longitude of the site at which the sound files were collected, expressed in decimal degrees.}

\item{twilight}{A character string of the twilight method to be used for sunrise and sunset as the boundary between day and night.
Options can be found in the \code{\link[photobiology]{day_night}} documentation.}

\item{dawnstart}{A numeric argument. If type='dawn', used to determine the start of dawn. By default, dawn starts at sunrise. Expressed as the time in seconds before sunrise.}

\item{dawnend}{A numeric argument. If type='dawn', used to determine the end of dawn. By default, dawn ends 1.5 hours after sunrise. Expressed as the time in seconds after sunrise.}

\item{duskstart}{A numeric argument. If type='dusk', used to determine the start of dusk. By default, dusk starts 1.5 hours before sunset. Expressed as the time in seconds before sunset.}

\item{duskend}{A numeric argument. If type='dusk', used to determine the end of dusk. By default, dusk ends at sunset. Expressed as the time in seconds after sunset.}

\item{output}{A character string. Indicates the format in which the soundscape diversity is expressed. Options are "percentage" (the fraction between the observed soundscape diversity and the maximum possible soundscape diversity), or "raw" (the number of acoustically active OSUs in the soundscape). Defaults to "percentage".}
}
\value{
A list of pairwise matrices for the beta diversity, and the one-complement for the Sorensen-type overlap, Jaccard-type overlap, Sorensen-type turnover and Jaccard-type turnover.
}
\description{
Computation of pairwise dissimularities among soundscapes.

Regular Beta diversity varies between 1-N, its value being dependent on the number of subsystems under consideration (N). As such, due to its dependence on the number of subsystems, it cannot directly be used as a measure of dissimilarity among communities. Instead, several simple transformations can be performed on the Beta values to get dissimilarity indices ranging between 0-1.


- Sorensen-type (local) overlap:

Quantifies the average proportion of a sub-system's OSUs which are shared across all considered sub-systems. It quantifies the overlap (similarity) from the sub-system perspective. To make this into a dissimilarity metric, we take the one-complement (1-Sorensen overlap), being the average proportion of non-shared OSUs in the system.

- Jaccard-type (regional) overlap:

Quantifies the effective proportion of OSUs which are shared across all subsystems. It quantifies overlap (similarity) from the perspective of the overall system. To make this into a dissimilarity metric, we take the one-complement (1 - Jaccard overlap), being the effective proportion of non-shard OSUs in the whole system.

- Sorensen-type turnover:

Quantifies the normalized turnover rate of OSUs from the perspective of the subsystem (alpha) - or the proportion of a subsystem which changes across subsystems. Once again, we take the one-complement as a dissimilarity measure.

- Jaccard-type turnover:

Quantifies the normalized OSU turnover rate from the perspective of the whole system (gamma). Once more, the one-complement gives us our dissimilarity measure.
}
