% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_function_names.R
\name{sounddiv_part}
\alias{sounddiv_part}
\title{sounddiv_part function}
\usage{
sounddiv_part(...)
}
\arguments{
\item{...}{no parameters}
}
\description{
Depreciated - Use 'ss_divpart' instead
}
