% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_function_names.R
\name{sounddiv_prof}
\alias{sounddiv_prof}
\title{sounddiv_prof function}
\usage{
sounddiv_prof(...)
}
\arguments{
\item{...}{no parameters}
}
\description{
Depreciated - Use 'ss_diversity_prof' instead
}
