% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundscape_diversity_calculations.R
\name{soundscape_diversity}
\alias{soundscape_diversity}
\title{Estimate Soundscape Diversity}
\usage{
soundscape_diversity(
  df,
  type = "total",
  date,
  lat,
  lon,
  minfreq = "default",
  maxfreq = "default",
  mintime = "default",
  maxtime = "default",
  twilight = "sunlight",
  dawnstart = 0,
  dawnend = 5400,
  duskstart = 5400,
  duskend = 0,
  freqseq = FALSE,
  nbins = 10,
  output = "percentage"
)
}
\arguments{
\item{df}{The aggregated time-frequency dataframe produced by \code{\link{aggregate_df}}.}

\item{type}{The scale for which the soundscape diversity is computed. Options are 'total', 'day',
'night', 'dawn', 'dusk' and 'tod' (time of day - for each unique time in the day).}

\item{date}{The first day of the recording period. Used for managing time-objects in R. \cr
Format as "YYYY-mm-dd".}

\item{lat}{The latitude of the site at which the sound files were collected.}

\item{lon}{The longitude of the site at which the sound files were collected.}

\item{minfreq}{The lower frequency limit for which to compute the soundscape diversity, expressed as a numeric value.
Defaults to the lowest frequency for which data exists in the dataframe.}

\item{maxfreq}{The upper frequency limit for which to compute the soundscape diversity, expressed as a numeric value.
Defaults to the highest frequency for which data exists in the dataframe.}

\item{mintime}{The lower time limit for which to compute the soundscape diversity, formatted as "HH:MM:SS". Defaults to the
earliest time for which data exists in the dataframe.}

\item{maxtime}{The upper time limit for which to compute the soundscape diversity, formatted as "HH:MM:SS". Defaults to the
latest time for which data exists in the dataframe.}

\item{twilight}{A character string of the twilight method to be used for sunrise and sunset as the boundary between day and night.
Options can be found in the \code{\link[photobiology]{day_night}} documentation.}

\item{dawnstart}{If type='dawn', used to determine the start of dawn. By default, dawn starts at sunrise. Expressed as the
time in seconds before sunrise.}

\item{dawnend}{If type='dawn', used to determine the end of dawn. By default, dawn ends 1.5 hours after sunrise. Expressed as the
time in seconds after sunrise.}

\item{duskstart}{If type='dusk', used to determine the start of dusk. By default, dusk starts 1.5 hours before sunset. Expressed as the
time in seconds before sunset.}

\item{duskend}{If type='dusk', used to determine the end of dusk. By default, dusk ends at sunset. Expressed as the
time in seconds after sunset.}

\item{freqseq}{One of either TRUE or FALSE. If set to FALSE, will compute the diversity for the entire frequency range of the soundscape.
If set to TRUE, will compute the diversity per frequency-bin of user-defined width.}

\item{nbins}{If freqseq is set to TRUE, determines the number of the frequency-bins by which divide the frequency range to compute the
soundscape diversity, expressed as a numeric constant.}

\item{output}{The format in which the soundscape diversity is expressed. Options are "percentage"
(the fraction between the observed soundscape diversity and the maximum possible soundscape diversity), or "raw"
(The raw Shannon-diversity of time-frequency bins). Defaults to "percentage".}
}
\value{
Depending on the chosen parameters, returns the soundscape diversity either a numeric value, a vector of values or a list of vectors of values.
}
\description{
Operates in a similar way as \code{\link{soundscape_richness}}. For a set acoustic index,
calculates the Shannon-diversity of time-frequency bins in the soundscape. In this case, each time-frequency
bin is treated as a "species", and the proportion of acoustically active recordings
for that bin the relative abundance. The soundscape diversity can be computed at various scales and resolutions.
For instance, the user can explore the diversity for the whole soundscape, specify custom time and frequency limits,
or use one of the built-in presets for diurnal-phase subsetting (day, night, dawn, dusk).
Additionally, the user can track the change in soundscape diversity throughout the day. Finally,
the diversity can be assessed for the entire frequency range, or per frequency-bin of user-defined width.

Note: the soundscape diversity metric should not be used to make inference about the diversity of the
real-world biological community unless verified using ground-truthing methods.
}
\keyword{internal}
