% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundscape_diversity_calculations.R
\name{soundscape_richness}
\alias{soundscape_richness}
\title{Estimate Soundscape Richness}
\usage{
soundscape_richness(
  df,
  type = "total",
  date,
  lat,
  lon,
  minfreq = "default",
  maxfreq = "default",
  mintime = "default",
  maxtime = "default",
  twilight = "sunlight",
  dawnstart = 0,
  dawnend = 5400,
  duskstart = 5400,
  duskend = 0,
  freqseq = FALSE,
  nbins = 10,
  output = "percentage"
)
}
\arguments{
\item{df}{The aggregated time-frequency dataframe produced by \code{\link{aggregate_df}}.}

\item{type}{The scale for which the soundscape richness is computed. Options are 'total', 'day',
'night', 'dawn', 'dusk' and 'tod' (time of day - for each unique time in the day).}

\item{date}{The first day of the recording period. Used for managing time-objects in R. \cr
Format as "YYYY-mm-dd".}

\item{lat}{The latitude of the site at which the sound files were collected.}

\item{lon}{The longitude of the site at which the sound files were collected.}

\item{minfreq}{The lower frequency limit for which to compute the soundscape richness, expressed as a numeric value.
Defaults to the lowest frequency for which data exists in the dataframe.}

\item{maxfreq}{The upper frequency limit for which to compute the soundscape richness, expressed as a numeric value.
Defaults to the highest frequency for which data exists in the dataframe.}

\item{mintime}{The lower time limit for which to compute the soundscape richness, formatted as "HH:MM:SS". Defaults to the
earliest time for which data exists in the dataframe.}

\item{maxtime}{The upper time limit for which to compute the soundscape richness, formatted as "HH:MM:SS". Defaults to the
latest time for which data exists in the dataframe.}

\item{twilight}{A character string of the twilight method to be used for sunrise and sunset as the boundary between day and night.
Options can be found in the \code{\link[photobiology]{day_night}} documentation.}

\item{dawnstart}{If type='dawn', used to determine the start of dawn. By default, dawn starts at sunrise. Expressed as the
time in seconds before sunrise.}

\item{dawnend}{If type='dawn', used to determine the end of dawn. By default, dawn ends 1.5 hours after sunrise. Expressed as the
time in seconds after sunrise.}

\item{duskstart}{If type='dusk', used to determine the start of dusk. By default, dusk starts 1.5 hours before sunset. Expressed as the
time in seconds before sunset.}

\item{duskend}{If type='dusk', used to determine the end of dusk. By default, dusk ends at sunset. Expressed as the
time in seconds after sunset.}

\item{freqseq}{One of either TRUE or FALSE. If set to FALSE, will compute the richness for the entire frequency range of the soundscape.
If set to TRUE, will compute the richness per frequency-bin of user-defined width.}

\item{nbins}{If freqseq is set to TRUE, determines the number of the frequency-bins by which to divide the frequency range to compute the
soundscape richness, expressed as a numeric constant.}

\item{output}{The format in which the soundscape richness is expressed. Options are "percentage"
(the fraction between the observed soundscape richness and the maximum possible soundscape richness), or "raw"
(the number of acoustically active time-frequency bins in the soundscape). Defaults to "percentage".}
}
\value{
Depending on the chosen parameters, returns the soundscape richness either a numeric value, a vector of values or a list of vectors of values.
}
\description{
For a set acoustic index, calculates the richness of acoustically active (value>0)
time-frequency bins in the soundscape, also expressed as the 'acoustic space use' or
'soundscape saturation'. The soundscape richness can be computed at various scales and resolutions.
For instance, the user can explore the richness for the whole soundscape, specify custom time and frequency limits,
or use one of the built-in presets for diurnal-phase subsetting (day, night, dawn, dusk).
Additionally, the user can track the change in soundscape richness throughout the day. Finally,
the richness can be assessed for the entire frequency range, or per frequency-bin of user-defined width.

Note: the soundscape richness metric should not be used to make inference about the richness of the
real-world biological community unless verified using ground-truthing methods.

For more information regarding the relationship between the soundscape richness and real-world diversity, consult:

Aide, T. M., Hernández-Serna, A., Campos-Cerqueira, M., Acevedo-Charry, O., & Deichmann, J. L. (2017). Species richness (of insects) drives the use of acoustic space in the tropics. Remote Sensing, 9(11), 1096.
\url{https://doi.org/10.3390/rs9111096}

Burivalova, Z., Towsey, M., Boucher, T., Truskinger, A., Apelis, C., Roe, P., & Game, E. T. (2018). Using soundscapes to detect variable degrees of human influence on tropical forests in Papua New Guinea. Conservation Biology, 32(1), 205-215.
\url{https://doi.org/10.1111/cobi.12968}
}
