% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_split_soundscape.R
\name{split_df}
\alias{split_df}
\title{Split soundscapes into 24h samples of acoustic trait space}
\usage{
split_df(binarized_soundscape, dailyfiles = 1440, maxdays = 10)
}
\arguments{
\item{binarized_soundscape}{The binarized soundscape object produced by the \code{\link{binarize_df}} function.}

\item{dailyfiles}{The number of files produced per 24h period the soundscape was sampled. This parameter depends on the sampling regime. The function defaults to 1440, which is the number of 1-min files produced per 24h period using a continuous sampling regime.}

\item{maxdays}{The maximum number of full 24h periods in the acoustic survey. Defaults to 10 days.}
}
\value{
Returns an updated soundscape object containing the binarized_df, but split per 24h sample of the acoustic trait space.
}
\description{
Split soundscapes into 24h samples of acoustic trait space
}
