% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_differential_heatmaps.R
\name{ss_compare}
\alias{ss_compare}
\title{Create differential soundscape heatmaps}
\usage{
ss_compare(
  aggregated_soundscape_A,
  aggregated_soundscape_B,
  type = "regular",
  timeinterval = "1 hour",
  mintime = "default",
  maxtime = "default",
  freqinterval = 2000,
  minfreq = 0,
  maxfreq = "default",
  labelsize_time = 4,
  labelsize_frequency = 4,
  labelsize_polar = 3,
  palette = "B",
  direction = 1,
  zero.black = FALSE,
  interactive = FALSE
)
}
\arguments{
\item{aggregated_soundscape_A}{An aggregated soundscape object produced by
\code{\link{ss_aggregate}} function for comparison.}

\item{aggregated_soundscape_B}{An aggregated soundscape object produced by
\code{\link{ss_aggregate}} function for comparison.}

\item{type}{One of either "regular" or "polar". If set
to "regular", produces a regular rectangular heatmap.
If set to "polar", produces a polar heatmap suitable for
 exploring diurnal patterns.}

\item{timeinterval}{A time interval for the x-axis.
Options can be found in the
\code{\link[scales]{breaks_width}} documentation.}

\item{mintime}{The lower time limit for the x-axis,
formatted as "HH:MM:SS". Defaults to the earliest
time for which data exists in the dataframe.}

\item{maxtime}{The upper time limit for the x-axis,
formatted as "HH:MM:SS".Defaults to the latest
time for which data exists in the dataframe.}

\item{freqinterval}{The frequency interval for the y-axis,
expressed as a numeric value.}

\item{minfreq}{The lower frequency limit for the y-axis
as a numeric value. Defaults to zero.}

\item{maxfreq}{The upper frequency limit for the y-axis
as a numeric value. Defaults to the maximum frequency of
 the dataframe.}

\item{labelsize_time}{If annotate=TRUE, can be used to
alter the size of temporal spectrum annotation. Please
 provide as a single positive integer with a value > 0.}

\item{labelsize_frequency}{If annotate=TRUE, can be used
to alter the size of frequency spectrum annotation.
Please provide as a single positive integer with a
value > 0.}

\item{labelsize_polar}{If type="polar", can be used to
alter the size of the frequency axis labels. Please
 provide as a single positive integer with a value > 0.}

\item{palette}{A character string indicating the
colormap option to use. Four options are available:
 "magma" (or "A"), "inferno" (or "B"), "plasma" (or "C"),
 "viridis" (or "D", the default option) and "cividis"
 (or "E"). Consult \url{https://www.rdocumentation.org/packages/viridisLite/versions/0.3.0/topics/viridis} for options.}

\item{direction}{Sets the order of colors in the scale.
If 1, the default, the regular order is followed.
If -1, the order of colors is reversed.}

\item{zero.black}{One of either TRUE or FALSE.
If set to TRUE, absent OSUs with incidence zero
will be colored black}

\item{interactive}{One of either TRUE or FALSE.
If set to TRUE, an interactive plot is produced using
 \code{\link[plotly]{ggplotly}}.}
}
\value{
Returns a differential heatmap
}
\description{
Creates a differential soundscape heatmap that shows the
difference in OSU presence and prevalence between two soundscape objects.
}
