% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_ss_create.R
\name{ss_create}
\alias{ss_create}
\title{Creating a soundscape object}
\usage{
ss_create(
  fileloc,
  samplerate,
  window,
  index = "CVR",
  date,
  lat,
  lon,
  method,
  value,
  output
)
}
\arguments{
\item{fileloc}{The full-length path to the directory where the
output of the \code{\link{ss_index_calc}} function is located.}

\item{samplerate}{The sampling rate specified in the \code{\link{ss_index_calc}}
function.}

\item{window}{The window length specified in the \code{\link{ss_index_calc}}
function.}

\item{index}{The acoustic index of interest. Options are
"BGN", "PMN", "CVR", "EVN", "ENT", "ACI", "OSC", "SPT", "RHZ",
"RVT", "RPS" and "RNG". For a brief description of indices, consult
the \code{\link{ss_index_calc}} documentation. Note that the soundscape diversity metrics
that can be calculated downstream have only been tested using the Acoustic Cover (CVR) index}

\item{date}{The first day of the recording period. Used for managing time-objects in R.
Formatted as "YYYY-mm-dd".}

\item{lat}{The latitude of the site at which the sound files were
collected. Coordinates should be specified in decimal degrees as a
 numerical variable.}

\item{lon}{The longitude of the site at which the sound files were
collected. Coordinates should be specified in decimal degrees as a
 numerical variable.}

\item{method}{The algorithm used to determine the threshold.
Options are "IJDefault","Huang", "Huang2", "Intermodes",
"IsoData", "Li", "MaxEntropy", "Mean", "MinErrorI", "Minimum",
"Moments", "Otsu", "Percentile", "RenyiEntropy", "Shanbhag",
"Triangle", "Yen", and "Mode". To specify a custom threshold,
use method="Custom" in combination with the value argument.
Consult \url{http://imagej.net/Auto_Threshold} for more
information on algorithm methodologies.}

\item{value}{Optional argument used to set a custom threshold
value for binarization - used in combination with method="Custom".}

\item{output}{Determines whether the function returns the raw
total number of detections per time during the
recording period (output = "raw"), or the incidence frequency
(total number of detections / number of recordings for that time
- output = "incidence_freq).}
}
\value{
Returns a soundscape object.
}
\description{
A wrapper function that combined the ss_index_merge, ss_binarize and ss_aggregate functions
to create a soundscape object in a single step
}
