% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_ss_divpart.R
\name{ss_divpart}
\alias{ss_divpart}
\title{Partitioning Soundscape Diversity into Alpha, Beta and Gamma components}
\usage{
ss_divpart(
  soundscape_list,
  qvalue,
  hier_table = NULL,
  minfreq = 0,
  maxfreq = "default",
  mintime = "default",
  maxtime = "default",
  output = "percentage"
)
}
\arguments{
\item{soundscape_list}{A list of soundscape objects of equal dimensions, each soundscape object being produced by \code{\link{ss_create}} (or \code{\link{ss_index_merge}}, \code{\link{ss_binarize}} and \code{\link{ss_aggregate}} in sequence).}

\item{qvalue}{A positive integer or decimal number (>=0), most commonly between 0-3. This parameter modulates the sensitivity of diversity values to the relative abundance of Operational Sound Units (OSUs). A value of 0 corresponds to the richness, a value of 1 is the equivalent number of effective OSUs for the Shannon index, a value of 2 is the equivalent number of effective OSUs for the Simpson index.}

\item{hier_table}{A matrix indicating the relationship between the soundscapes in the soundscape_list. The first column lists the names of all the soundscapes in the soundscape_list, other columns can be used to group soundscapes into higher hierarchical levels. If no hierarchy table is supplied, the function defaults to a 2-level diversity partitioning.}

\item{minfreq}{A numeric value indicating the lower frequency limit for which to compute the soundscape diversity. If set to default, uses the lowest available frequency in the dataframe.}

\item{maxfreq}{A numeric value indicating the upper frequency limit for which to compute the soundscape diversity. If set to default, uses the highest available frequency in the dataframe.}

\item{mintime}{The lower time limit for which to compute the soundscape diversity, formatted as "HH:MM:SS". If set to default, uses the earliest time for which data exists in the dataframe.}

\item{maxtime}{The upper time limit for which to compute the soundscape diversity, formatted as "HH:MM:SS". If set to default, uses the latest time for which data exists in the dataframe.}

\item{output}{A character string. Indicates the format in which the soundscape diversity is expressed. Options are "percentage" (the fraction between the observed soundscape diversity and the maximum possible soundscape diversity), or "raw" (the number of acoustically active OSUs in the soundscape). Defaults to "percentage".}
}
\value{
A dataframe of diversity values
}
\description{
The diversity of a system can generally be broken down into three components:
Alpha, Beta and Gamma diversity.

For Hill numbers these three components take a multiplicative relationship: Gamma = Alpha x Beta

- \bold{Alpha diversity}: The average diversity of the subsystems

- \bold{Beta diversity}: The compositional heterogeneity between subsystems.

The beta diversity component can be computed as: Beta = Gamma / Alpha.

Beta represents the effective number of equally large and completely unique subsystems within the system. As Beta quantifies the ratio between Gamma and Alpha, it can also be seen as the number of times more diverse the whole system is in effective number of OSUs compared to its constituent subsystems on average. The beta diversity ranges from 1 to N (the number of subsystems in the system).

- \bold{Gamma diversity}: The average diversity of the whole system
}
