% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_ss_diversity.R
\name{ss_evenness}
\alias{ss_evenness}
\title{Estimate the soundscape evenness using Hill numbers}
\usage{
ss_evenness(
  soundscape_obj = soundscape_obj,
  subset = "total",
  mintime = "default",
  maxtime = "default",
  minfreq = 0,
  maxfreq = "default",
  dawnstart = 0,
  dawnend = 5400,
  duskstart = 5400,
  duskend = 0
)
}
\arguments{
\item{soundscape_obj}{The aggregated soundscape object produced by
\code{\link{ss_aggregate}} function.}

\item{subset}{The scale for which the soundscape diversity is computed.
Options are 'total', 'day', night', 'dawn', 'dusk' and
'tod' (time of day - for each unique time in the day).}

\item{mintime}{The lower time limit for which to compute the soundscape
diversity, formatted as "HH:MM:SS". If set to default, uses the
earliest time for which data exists in the dataframe.}

\item{maxtime}{The upper time limit for which to compute the soundscape
diversity, formatted as "HH:MM:SS". If set to default, uses the
latest time for which data exists in the dataframe.}

\item{minfreq}{A numeric value indicating the lower frequency limit
for which to compute the soundscape diversity. If set to default, uses
the lowest available frequency in the dataframe.}

\item{maxfreq}{A numeric value indicating the upper frequency limit
for which to compute the soundscape diversity. If set to default,
uses the highest available frequency in the dataframe.}

\item{dawnstart}{A numeric argument. If subset == 'dawn', used to determine
the start of dawn. By default, dawn starts at sunrise. Expressed as
the time in seconds before sunrise.}

\item{dawnend}{A numeric argument. If subset == 'dawn', used to determine
the end of dawn. By default, dawn ends 1.5 hours after sunrise.
Expressed as the time in seconds after sunrise.}

\item{duskstart}{A numeric argument. If subset == 'dusk', used to determine
the start of dusk. By default, dusk starts 1.5 hours before sunset.
 Expressed as the time in seconds before sunset.}

\item{duskend}{A numeric argument. If subset == 'dusk', used to determine the
end of dusk. By default, dusk ends at sunset. Expressed as the
time in seconds after sunset.}
}
\value{
Depending on the chosen parameters, returns the soundscape
diversity either a numeric value, a vector of values or a list of
vectors of values.
}
\description{
Calculated the soundscape evenness index for the provided soundscape object.
The user can explore the diversity for the whole soundscape, specify custom
time and frequency limits, or use one of the built-in presets for
diurnal-phase subsetting (day, night, dawn, dusk). Additionally, the user
can track the change in soundscape diversity throughout the day.

\strong{Note:} Soundscape diversity metrics should not be used to make
 inference about the diversity of the real-world biological community
  unless verified using ground-truthing methods.

\strong{Note:} In \href{https://besjournals.onlinelibrary.wiley.com/doi/full/10.1111/2041-210X.13924}{Luypaert et al. (2022)},
we stated that the soundscape evenness index should be calculated as:
soundscape diversity (2D) / soundscape richness (0D).

However, based on recent publications, a more proper way to calculate the
soundscape evenness is:
(soundscape diversity (2D) - 1) / (soundscape richness (0D) - 1).

Therefore, from now on, we will use the latter equation to calculate the soundscape evenness index.
}
