% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_ss_index_calc.R
\name{ss_find_files}
\alias{ss_find_files}
\title{Find all .wav files in a parent directory}
\usage{
ss_find_files(parent_directory)
}
\arguments{
\item{parent_directory}{The full-length path to a directory containing .wav files, or a directory containing
directories with .wav files. If the latter, this function will look for .wav files recursively in all subdirectories.}
}
\value{
A named list with full-length paths to the '.wav' files contained in the parent directory or subdirectories
}
\description{
This function takes the path to a parent directory and recursively looks for all directories and subdirectories
containing .wav files. The full-length path to each discovered .wav files is saved in a named list maintaining
directory structure.
}
\examples{

# Get file path to '.wav' files
fpath <- system.file("extdata", package = "soundscapeR")

# Finding files
file_locs <- ss_find_files(parent_directory = fpath)

}
