% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_ss_binarize.R
\name{ss_get_mode}
\alias{ss_get_mode}
\title{Calculate the mode}
\usage{
ss_get_mode(df)
}
\arguments{
\item{df}{The merged_soundscape data frame produced by
\code{\link{ss_index_merge}}.}
}
\value{
Returns the acoustic index value which appears most often
}
\description{
Calculate the mode
}
\details{
Function for internal use by \code{\link{ss_threshold}}
}
