% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_ss_index_calc.R, R/1_ss_index_calc_2.R
\name{ss_index_calc}
\alias{ss_index_calc}
\title{Compute Acoustic Indices}
\usage{
ss_index_calc(
  file_list,
  output_dir = NA,
  window = 256,
  theta = 3,
  parallel = FALSE
)

ss_index_calc(
  file_list,
  output_dir = NA,
  window = 256,
  theta = 3,
  parallel = FALSE
)
}
\arguments{
\item{file_list}{A list containing the full-length path to each '.wav' file in a folder. This list can be obtained using the
`ss_find_files` and `ss_assess_files` functions.}

\item{output_dir}{The full-length path to an output directory on your device. If not specified, will default to the location where
the raw sound files are saved.}

\item{window}{The window length used for the Fast-Fourier Transformation. Defaults to 256}

\item{theta}{The cut-off dB-value above which sound is considered present in a frequency bin. Defaults to 3 dB.}

\item{parallel}{A boolean operator indicating whether parallel processing should be used to calculate the CVR-index files.
Defaults to FALSE.}

\item{fileloc}{The full-length path to the sound file for which to
compute indices.}

\item{outputloc}{The full-length path to the location where you wish to save the output files.
Defaults to the same location as the fileloc.}

\item{samplerate}{The number of times the sound was sampled each second.
This is a fixed parameter determined by your recording setup, although
downsampling to a lower sampling rate is possible.}
}
\value{
Creates a directory with '.csv' files for the aforementioned
set of spectral and summary indices. The directory is created in the same
place as the original sound file.

A list of CVR-values for each sound file
}
\description{
Calls on the external software 'AnalysisPrograms.exe'
to compute a set of acoustic indices using user-specified
configuration parameters.

For a given folder of sound files, this function is used to calculate the CVR-index for each 1-min file.
}
\details{
The default duration of sound files for index computation is 60 seconds.
If the file length exceeds 60 seconds, the file is automatically cut into
60 second segments for further analysis.

\bold{The following spectral indices will be computed:}

\emph{Background Noise (BGN)}:
The mode of the distribution of decibel values in each
frequency bin, representing the background intensity value.
This index captures the acoustic energy which persists throughout the
duration of the sound file, regardless of its origin
(biophonic, geophonic or anthrophonic).

\emph{Power minus noise (PMN)}:
The difference between the maximum decibel value in
each frequency bin and the corresponding BGN decibel value.

\emph{Acoustic cover (CVR)}:
The fraction of active elements in each noise-reduced
frequency bin where the amplitude exceeds a 3-dB threshold.

\emph{Number of events (EVN)}:
The number of times the decibel value in a noise-reduced
frequency bin crosses the 3-dB threshold from lower to higher values.

\emph{Temporal entropy (ENT)}:
A measure of acoustic energy concentration in each noise-reduced
frequency bin.

\emph{Acoustic Complexity Index (ACI)}:
A measure quantifying the variability in
intensity values in each noise-reduced frequency bin. It is widely used as a
measure of biophony in recordings, however remains highly sensitive to
non-biological sources of sound.

\emph{Oscillation Index (OSC)}:
...

\emph{Spectral Peak Tracks (SPT)}:
A measure of the presence of spectral peak
tracks in a noise-reduced frequency bin.

\emph{Ridge indices (RHZ, RVT, RPS, RNS)}:
A set of indices based on the presence of formants
in the harmonic structure of many animal vocalizations, calculated in the
four directions of the ridge slope (horizontal (RHZ), vertical (RVT),
upward slope (RPS), downward slope (RNS)). Formants in the mid-band are
typically due to birdsong, whereas vertical formants are typical for
non-biological sounds such as rain drops and electrical clicks.

For index computation this function calls on the external software
'AnalysisPrograms.exe' by the QUT Ecoacoustics Research Group
(\url{https://github.com/QutEcoacoustics/audio-analysis}).Make sure the
software is installed prior to commencing the analysis. Additionally, the
software requires sound files to be named according to certain standards
(\url{https://ap.qut.ecoacoustics.info/basics/dates.html}).

\strong{Software citation}:
Michael Towsey, Anthony Truskinger, Mark Cottman-Fields, & Paul Roe.
(2018, March 5). Ecoacoustics Audio Analysis Software v18.03.0.41
(Version v18.03.0.41). Zenodo. http://doi.org/10.5281/zenodo.1188744
}
