% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_ss_index_calc.R
\name{ss_index_config}
\alias{ss_index_config}
\title{Set Configuration Parameters for Acoustic Index Computation}
\usage{
ss_index_config(samplerate = 41000, window = 256)
}
\arguments{
\item{samplerate}{The number of times the sound was sampled each second.
This is a fixed parameter determined by your recording setup, although
 downsampling to a lower sampling rate is possible.}

\item{window}{A variable of the Short-time Fourier Transformation,
expressed as the number of samples. The window size of choice
depends on the fundamental frequency, intensity and change of the signal
of interest, and influences the temporal and frequency resolution
of the analysis. The window size is generally a power of 2.}
}
\value{
Edits and saves the '.yml' configuration file which is
used by the AnalysisPrograms software for index computation.
}
\description{
Alters and saves the configuration file used by
external software 'AnalysisPrograms.exe' for the computation of
acoustic indices.
}
\details{
The \code{ss_index_config()} function is normally only used internally
by the \code{ss_index_calc()} function, although it is possible for the
user to edit the configuration file manually.
}
