% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ss_index_merge.R
\name{ss_index_merge}
\alias{ss_index_merge}
\title{Import and Merge Acoustic Index '.csv' Files}
\usage{
ss_index_merge(fileloc, samplerate, window, index, date, lat, lon)
}
\arguments{
\item{fileloc}{The full-length path to the directory where the
output directory of the \code{\link{ss_index_calc}} function are located.}

\item{samplerate}{The sampling rate specified in the \code{\link{ss_index_calc}}
function.}

\item{window}{The window length specified in the \code{\link{ss_index_calc}}
function.}

\item{index}{The acoustic index of interest. Options are
"BGN", "PMN", "CVR", "EVN", "ENT", "ACI", "OSC", "SPT", "RHZ",
"RVT", "RPS" and "RNG". For a brief description of indices, consult
the \code{\link{ss_index_calc}} documentation. For the "BGN" index, the minimum index
 value was added to all data frame values to obtain a data frame of positive
 values while retaining the relationship between individual OSUs}

\item{date}{The first day of the recording period. Used
for managing time-objects in R.
Formatted as "YYYY-mm-dd".}

\item{lat}{The latitude of the site at which the sound files were
collected. Coordinates should be specified in decimal degrees as a
 numerical variable.}

\item{lon}{The longitude of the site at which the sound files were
collected. Coordinates should be specified in decimal degrees as a
 numerical variable.}
}
\value{
Returns a time-by-frequency dataframe of acoustic index values
for all files in the recording period.
}
\description{
For an acoustic index of choice, imports all
spectral index '.csv' files produced by the \code{\link{ss_index_calc}}
function, and merges them into a time-by-frequency data frame.
}
