% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_ss_divpart.R
\name{ss_pairdis}
\alias{ss_pairdis}
\title{Pairwise Beta diversity and dissimilarity values between subsystems in the system}
\usage{
ss_pairdis(
  soundscape_list,
  qvalue = 0,
  hier_table = NA,
  minfreq = 0,
  maxfreq = "default",
  mintime = "default",
  maxtime = "default"
)
}
\arguments{
\item{soundscape_list}{A list of dataframes of equal dimensions, each dataframe being produced by \code{\link{ss_aggregate}}.}

\item{qvalue}{A positive integer or decimal number (>=0), most commonly between 0-3. This parameter modulates the sensitivity of diversity values to the relative abundance of Operational Sound Units (OSUs). A value of 0 corresponds to the richness, a value of 1 is the equivalent number of effective OSUs for the Shannon index, a value of 2 is the equivalent number of effective OSUs for the Simpson index.}

\item{hier_table}{A matrix indicating the relationship between the soundscapes in the soundscape_list. The first column lists the names of all the soundscapes in the soundscape_list, other columns can be used to group soundscapes into higher hierarchical levels. If no hierarchy table is supplied, the function defaults to a 2-level diversity partitioning.}

\item{minfreq}{A numeric value indicating the lower frequency limit for which to compute the soundscape diversity. If set to default, uses the lowest available frequency in the dataframe.}

\item{maxfreq}{A numeric value indicating the upper frequency limit for which to compute the soundscape diversity. If set to default, uses the highest available frequency in the dataframe.}

\item{mintime}{The lower time limit for which to compute the soundscape diversity, formatted as "HH:MM:SS". If set to default, uses the earliest time for which data exists in the dataframe.}

\item{maxtime}{The upper time limit for which to compute the soundscape diversity, formatted as "HH:MM:SS". If set to default, uses the latest time for which data exists in the dataframe.}
}
\value{
A list of pairwise matrices for the beta diversity, and the one-complement for the Sorensen-subset overlap, Jaccard-subset overlap, Sorensen-subset turnover and Jaccard-subset turnover.
}
\description{
Computation of pairwise dissimularities among soundscapes.

Regular Beta diversity varies between 1-N, its value being dependent on the number of subsystems under consideration (N). As such, due to its dependence on the number of subsystems, it cannot directly be used as a measure of dissimilarity among communities. Instead, several simple transformations can be performed on the Beta values to get dissimilarity indices ranging between 0-1.


- Sorensen-subset (local) overlap:

Quantifies the average proportion of a sub-system's OSUs which are shared across all considered sub-systems. It quantifies the overlap (similarity) from the sub-system perspective. To make this into a dissimilarity metric, we take the one-complement (1-Sorensen overlap), being the average proportion of non-shared OSUs in the system.

- Jaccard-subset (regional) overlap:

Quantifies the effective proportion of OSUs which are shared across all subsystems. It quantifies overlap (similarity) from the perspective of the overall system. To make this into a dissimilarity metric, we take the one-complement (1 - Jaccard overlap), being the effective proportion of non-shard OSUs in the whole system.

- Sorensen-subset turnover:

Quantifies the normalized turnover rate of OSUs from the perspective of the subsystem (alpha) - or the proportion of a subsystem which changes across subsystems. Once again, we take the one-complement as a dissimilarity measure.

- Jaccard-subset turnover:

Quantifies the normalized OSU turnover rate from the perspective of the whole system (gamma). Once more, the one-complement gives us our dissimilarity measure.
}
