% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_principle_coordinate_plots.R
\name{ss_pcoa}
\alias{ss_pcoa}
\title{Principle Coordinate plots for a list of soundscapes}
\usage{
ss_pcoa(aggregated_soundscape_list, grouping = NA, screeplot = FALSE)
}
\arguments{
\item{aggregated_soundscape_list}{A list of aggregated_soundscape objects. Each object in the
list should be produced using the \code{\link{ss_aggregate}} function.}

\item{grouping}{A numeric or character vector indicating potential grouping of the elements in the 'aggregated_soundscape_list' object.
Make sure that the grouping vector has the same length as the 'aggregated_soundscape_list' argument.}

\item{screeplot}{A boolean operator indicating whether the function should produce a screeplot in addition to the PCoA plot.
Options are: TRUE and FALSE.}
}
\value{
Returns a principal coordinate plot
}
\description{
Creates a Principle Coordinate plot, using a dimensionality reduction
approach to plot the distances between soundscapes in 2D space based on the Bray-Curtis dissimilarity.
}
