% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_ss_index_calc.R
\name{ss_split_files}
\alias{ss_split_files}
\title{Check if a file is the correct length, and if not, split it into 60-second chunks.}
\usage{
ss_split_files(file_locs)
}
\arguments{
\item{file_locs}{The output of the `ss_find_files` function}
}
\value{
In case some files are longer than 60 seconds, creates a new directory called 'split_files' containing the sound files split into 60-second chunks.
}
\description{
This function takes the output list of the `ss_find_files` function and checks whether the sound files in the folder
are the correct length (60 seconds). If not, the function splits the sound files into 60-second chunks and saves the files in a new
sub-directory called 'split_files'.
}
