% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_ss_binarize.R
\name{ss_threshold}
\alias{ss_threshold}
\title{Determine Binarization Threshold}
\usage{
ss_threshold(merged_soundscape, method)
}
\arguments{
\item{merged_soundscape}{The merged_soundscape object produced by
\code{\link{ss_index_merge}}.}

\item{method}{The algorithm used to determine the threshold.
Options are "IJDefault", "Huang", "Huang2", "Intermodes",
"IsoData", "Li", "MaxEntropy", "Mean", "MinErrorI", "Minimum",
"Moments", "Otsu", "Percentile", "RenyiEntropy", "Shanbhag",
"Triangle", "Yen", and "Mode".
Consult \url{http://imagej.net/Auto_Threshold} for more
information on algorithm methodologies.}
}
\value{
Returns a numeric threshold value for subsequent
binarization.
}
\description{
Determines the threshold for binarization of the
time-frequency dataframe of index values. Several binarization
algorithms are available, either based on image thresholding
algorithms from \code{\link[autothresholdr]{auto_thresh}}, or
 calculation of the mode.
}
\details{
Function for internal use by \code{\link{ss_binarize}},
however can also be called on by the user manually.
}
