% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_ss_binarize.R
\name{ss_threshold_check}
\alias{ss_threshold_check}
\title{Check which binarization algorithm works for a set index}
\usage{
ss_threshold_check(merged_soundscape, method, value = NULL)
}
\arguments{
\item{merged_soundscape}{The merged_soundscape object produced by
\code{\link{ss_index_merge}}.}

\item{method}{The algorithm used to determine the
threshold. Options are "IJDefault","Huang", "Huang2",
 "Intermodes", "IsoData", "Li", "MaxEntropy", "Mean",
 "MinErrorI", "Minimum", "Moments", "Otsu", "Percentile",
 "RenyiEntropy", "Shanbhag", "Triangle", "Yen", and
 "Mode". To specify a custom threshold, use
 method="Custom" in combination with the value argument.
 Consult \url{http://imagej.net/Auto_Threshold} for more
 information on algorithm methodologies.}

\item{value}{Optional argument used to set a custom
threshold value for binarization - used in combination
 with method="Custom".}
}
\value{
A plot with before and after binarization to
 facilitate visual comparison of binarization algorithm
  performance.
}
\description{
Plots the merged acoustic index dataframe
 before and after the binarization step using a
 user-specified binarization algorithm. Facilitates the
 rapid visual comparison of binarization performance using
 different algorithms.
}
