% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_threshold_aggregate.R
\name{threshold_df}
\alias{threshold_df}
\title{Determine Binarization Threshold}
\usage{
threshold_df(df, method)
}
\arguments{
\item{df}{The time-frequency dataframe of index values produced by \code{merge_csv}.}

\item{method}{The algorithm used to determine the threshold. Options are "IJDefault",
"Huang", "Huang2", "Intermodes", "IsoData", "Li", "MaxEntropy", "Mean", "MinErrorI", "Minimum",
"Moments", "Otsu", "Percentile", "RenyiEntropy", "Shanbhag", "Triangle", "Yen", and "Mode.
Consult \url{http://imagej.net/Auto_Threshold} for more information on algorithm methodologies.}
}
\value{
Returns a numeric threshold value for subsequent binarization.
}
\description{
Determines the threshold for binarization of the
time-frequency dataframe of index values. Several binarization algorithms are available, either based on image thresholding
algorithms from \code{\link[autothresholdr]{auto_thresh}}, or calculation of the mode.
}
\details{
Function for internal use by \code{\link{binarize_df}}, however can also be called on by the user manually.
}
