% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_aggregate_df.R
\name{aggregate_df}
\alias{aggregate_df}
\title{Aggregate binarized dataframe}
\usage{
aggregate_df(binarized_soundscape, output = "incidence_freq")
}
\arguments{
\item{binarized_soundscape}{The binarized soundscape object produced by
\code{\link{binarize_df}}.}

\item{output}{Determines whether the function returns the raw
total number of detections (activity = 1) per time during the
recording period (output = "raw"), or the incidence frequency
(total number of detections / number of recordings for that time
- output = "incidence_freq).}
}
\value{
Returns a list containing three elements:
(i) aggregated_per_time: a list of soundscape samples
(the columns in the binarized_df function) grouped per
 unique time in the recording period; (
 ii) sampling_effort_per_time: a list indicating the
  sampling effort (number of soundscape samples) per
   unique time in the recording period;
   (iii) aggregated_df: an aggregated time-frequency
   data frame containing the number of acoustically
   active recordings in each bin for a set acoustic index.
}
\description{
Aggregates binarized spectral indices by time of day.
For each aggregate time period, the activity values (0 or 1)
are summed per frequency bin, and potentially divided by the number
of recordings for that time to get the proportion of
acoustically active recordings in each time-frequency bin
(the incidence frequency).
}
