% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_function_names.R
\name{aggregate_df}
\alias{aggregate_df}
\title{aggregate_df function}
\usage{
aggregate_df(...)
}
\arguments{
\item{...}{no parameters}
}
\description{
Depreciated - Use 'ss_aggregate' instead
}
