% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_function_names.R
\name{check_thresh}
\alias{check_thresh}
\title{check_thresh function}
\usage{
check_thresh(...)
}
\arguments{
\item{...}{no parameters}
}
\description{
Depreciated - Use 'ss_threshold_check' instead
}
