% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_data.R
\name{check_thresh}
\alias{check_thresh}
\title{Rapid soundscape heatmapping function to assess the performance of the thresholding step}
\usage{
check_thresh(df, method, value = NULL, strictness = 1)
}
\arguments{
\item{df}{A time-frequency dataframe produced by \code{\link{merge_csv}}.}

\item{method}{The algorithm used to determine the threshold. Options are "IJDefault",
"Huang", "Huang2", "Intermodes", "IsoData", "Li", "MaxEntropy", "Mean", "MinErrorI", "Minimum",
"Moments", "Otsu", "Percentile", "RenyiEntropy", "Shanbhag", "Triangle", "Yen", and "Mode.
Consult \url{http://imagej.net/Auto_Threshold} for more information on algorithm methodologies.}

\item{value}{Optional argument used to set a custom threshold value for binarization - used in combination with
method="Custom".}

\item{strictness}{A numeric value with which the threshold value is multiplied. Used for slight tweaking of
the threshold value. Reduce the threshold value using values <0, increase using values >1. Defaults to 1.}
}
\value{
Returns a before-after thresholding heatmap plot.
}
\description{
A quick-and-dirty heatmap for rapid data assessment of the performance of the thresholding step, using the \code{\link[plsgenomics]{matrix.heatmap}} function.
Useful for checking which binarization algorithm works best, and fine-tuning threshold values.
}
