% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_threshold_aggregate.R
\name{get_mode}
\alias{get_mode}
\title{Calculate the mode}
\usage{
get_mode(df)
}
\arguments{
\item{df}{The time-frequency dataframe of index values produced by \code{merge_csv}}
}
\value{
Returns the acoustic index value which appears most often
}
\description{
Calculate the mode
}
\details{
Function for internal use by \code{\link{threshold_df}}
}
