% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_function_names.R
\name{index_config}
\alias{index_config}
\title{index_config function}
\usage{
index_config(...)
}
\arguments{
\item{...}{no parameters}
}
\description{
Depreciated - Use 'ss_index_config' instead
}
