% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_diversity_metrics.R
\name{sounddiv_by_time}
\alias{sounddiv_by_time}
\title{Visualize Soundscape Diversity by Time of Day}
\usage{
sounddiv_by_time(
  df,
  qvalue,
  graphtype = "total",
  date,
  lat,
  lon,
  minfreq = "default",
  maxfreq = "default",
  nbins = 10,
  timeinterval = "1 hour",
  smooth = TRUE,
  movavg = 6,
  interactive = FALSE,
  save = FALSE,
  dir = "default",
  filename = "file",
  device = "png",
  output = "percentage",
  width = 150,
  height = 150
)
}
\arguments{
\item{df}{The aggregated time-frequency dataframe produced by \code{\link{aggregate_df}}.}

\item{qvalue}{A positive integer or decimal number (>=0), most commonly between 0-3. This parameter modulates the sensitivity of diversity values to the relative abundance of Operational Sound Units (OSUs). A value of 0 corresponds to the richness, a value of 1 is the equivalent number of effective OSUs for the Shannon index, a value of 2 is the equivalent number of effective OSUs for the Simpson index.}

\item{graphtype}{The type of plot which is produced.

Options are:

\emph{'total'}:

An area chart showing the soundscape diversity by time-of-day for the entire frequency range.

\emph{'frequency'}:

A stacked area chart showing the relative contribution of frequency bins with user-defined width to the total soundscape diversity by time-of-day.

\emph{'normfreq'}:

A percentage stacked area chart showing the normalized relative contribution of frequency bins with user-defined width to the soundscape diversity by time-of-day.

\emph{'linefreq'}:

A line chart showing the relative contribution of frequency bins with user-defined width to the soundscape diversity by time-of-day.}

\item{date}{The first day of the recording period. Used for managing time-objects in R.
Format as "YYYY-mm-dd".}

\item{lat}{The latitude of the site at which the sound files were collected.}

\item{lon}{The longitude of the site at which the sound files were collected.}

\item{minfreq}{The lower frequency limit for which to visualize the soundscape diversity, expressed as a numeric value.
Defaults to the lowest frequency for which data exists in the dataframe.}

\item{maxfreq}{The upper frequency limit for which to visualize the soundscape diversity, expressed as a numeric value.
Defaults to the highest frequency for which data exists in the dataframe.}

\item{nbins}{If graphtype='frequency'/'normfreq'/'linefreq', determines the number of the frequency-bins by which to divide the frequency range to compute the
relative contribution of each bin to the total diversity.}

\item{timeinterval}{A time interval for the x-axis. Options can be found in the \code{\link[scales]{date_breaks}} documentation.}

\item{smooth}{One of either TRUE or FALSE. If set to TRUE, applies a moving average filter for smoothing the diversity by time-of-day.}

\item{movavg}{If smooth=TRUE, determines the width of the moving average filter. Consult \code{\link[pracma]{movavg}} for more information.}

\item{interactive}{One of either TRUE or FALSE. If set to TRUE, an interactive plot is produced using \code{\link[plotly]{ggplotly}}.}

\item{save}{One of either TRUE or FALSE. If set to TRUE, saves the plot using \code{\link[ggplot2]{ggsave}}, and the 'dir', 'filename' and 'device'
arguments.}

\item{dir}{Path of the directory to save plot to: path and filename are combined to create the fully qualified file name.
Defaults to the working directory. For more information consult \code{\link[ggplot2]{ggsave}}.}

\item{filename}{The file name without the extention. For more information consult \code{\link[ggplot2]{ggsave}}.}

\item{device}{Device to use. Can either be a device function (e.g. png()), or one of
"eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).
Defaults to "png". For more information consult \code{\link[ggplot2]{ggsave}}.}

\item{output}{A character string. Indicates the format in which the soundscape diversity is expressed. Options are "percentage" (the fraction between the observed soundscape diversity and the maximum possible soundscape diversity), or "raw" (the number of acoustically active OSUs in the soundscape). Defaults to "percentage".}

\item{width}{If save=TRUE, expresses the width of the saved image in milimeters. Defaults to 100 mm.}

\item{height}{If save=TRUE, expresses the height of the saved image in milimeters. Defaults to 100 mm.}
}
\value{
Returns a ggplot object and if save=TRUE, saves the plot in a directory of choice using a specified device and filename.
}
\description{
Produces plots showing the variation in soundscape diversity by time-of-day. Soundscape diversity can be shown for the full frequency range, or the relative contribution of frequency-bins with user-specified width.
}
