% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_function_names.R
\name{sounddiv_pairdis}
\alias{sounddiv_pairdis}
\title{sounddiv_pairdis function}
\usage{
sounddiv_pairdis(...)
}
\arguments{
\item{...}{no parameters}
}
\description{
Depreciated - Use 'ss_pairdis' instead
}
