% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_partitioning.R
\name{sounddiv_part}
\alias{sounddiv_part}
\title{Partitioning Soundscape Diversity into Alpha, Beta and Gamma components}
\usage{
sounddiv_part(
  df_list,
  qvalue,
  hier_table = "default",
  type = "total",
  minfreq = "default",
  maxfreq = "default",
  mintime = "default",
  maxtime = "default",
  date,
  lat,
  lon,
  twilight = "sunlight",
  dawnstart = 0,
  dawnend = 5400,
  duskstart = 5400,
  duskend = 0,
  output = "percentage"
)
}
\arguments{
\item{df_list}{A list of dataframes of equal dimensions, each dataframe being produced by \code{\link{aggregate_df}}.}

\item{qvalue}{A positive integer or decimal number (>=0), most commonly between 0-3. This parameter modulates the sensitivity of diversity values to the relative abundance of Operational Sound Units (OSUs). A value of 0 corresponds to the richness, a value of 1 is the equivalent number of effective OSUs for the Shannon index, a value of 2 is the equivalent number of effective OSUs for the Simpson index.}

\item{hier_table}{A matrix indicating the relationship between the soundscapes in the df_list. The first column lists the names of all the soundscapes in the df_list, other columns can be used to group soundscapes into higher hierarchical levels. If no hierarchy table is supplied, the function defaults to a 2-level diversity partitioning.}

\item{type}{The scale for which the soundscape diversity is computed. Options are 'total', 'day',
'night', 'dawn', 'dusk' and 'tod' (time of day - for each unique time in the day).}

\item{minfreq}{A numeric value indicating the lower frequency limit for which to compute the soundscape diversity. If set to default, uses the lowest available frequency in the dataframe.}

\item{maxfreq}{A numeric value indicating the upper frequency limit for which to compute the soundscape diversity. If set to default, uses the highest available frequency in the dataframe.}

\item{mintime}{The lower time limit for which to compute the soundscape diversity, formatted as "HH:MM:SS". If set to default, uses the earliest time for which data exists in the dataframe.}

\item{maxtime}{The upper time limit for which to compute the soundscape diversity, formatted as "HH:MM:SS". If set to default, uses the latest time for which data exists in the dataframe.}

\item{date}{The first day of the recording period. Used for managing time-objects in R. \cr
Format as "YYYY-mm-dd".}

\item{lat}{The latitude of the site at which the sound files were collected, expressed in decimal degrees.}

\item{lon}{The longitude of the site at which the sound files were collected, expressed in decimal degrees.}

\item{twilight}{A character string of the twilight method to be used for sunrise and sunset as the boundary between day and night.
Options can be found in the \code{\link[photobiology]{day_night}} documentation.}

\item{dawnstart}{A numeric argument. If type='dawn', used to determine the start of dawn. By default, dawn starts at sunrise. Expressed as the time in seconds before sunrise.}

\item{dawnend}{A numeric argument. If type='dawn', used to determine the end of dawn. By default, dawn ends 1.5 hours after sunrise. Expressed as the time in seconds after sunrise.}

\item{duskstart}{A numeric argument. If type='dusk', used to determine the start of dusk. By default, dusk starts 1.5 hours before sunset. Expressed as the time in seconds before sunset.}

\item{duskend}{A numeric argument. If type='dusk', used to determine the end of dusk. By default, dusk ends at sunset. Expressed as the time in seconds after sunset.}

\item{output}{A character string. Indicates the format in which the soundscape diversity is expressed. Options are "percentage" (the fraction between the observed soundscape diversity and the maximum possible soundscape diversity), or "raw" (the number of acoustically active OSUs in the soundscape). Defaults to "percentage".}
}
\value{
A dataframe of diversity values
}
\description{
The diversity of a system can generally be broken down into three components: \cr
Alpha, Beta and Gamma diversity.\cr
\cr
For Hill numbers these three components take a multiplicative relationship: Gamma = Alpha x Beta \cr
\cr
- \bold{Alpha diversity}: The average diversity of the subsystems \cr
\cr
- \bold{Beta diversity}: The compositional heterogeneity between subsystems.\cr
\cr
The beta diversity component can be computed as: Beta = Gamma / Alpha. \cr
\cr
Beta represents the effective number of equally large and completely unique subsystems within the system. As Beta quantifies the ratio between Gamma and Alpha, it can also be seen as the number of times more diverse the whole system is in effective number of OSUs compared to its constituent subsystems on average. The beta diversity ranges from 1 to N (the number of subsystems in the system).\cr
\cr
- \bold{Gamma diversity}: The average diversity of the whole system \cr
}
