% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundscape_diversity_profiles.R
\name{sounddiv_prof}
\alias{sounddiv_prof}
\title{Compute and Visualize Soundscape Diversity Profiles}
\usage{
sounddiv_prof(
  aggregate_list,
  qvalues = seq(0, 5, 0.1),
  subset = "total",
  minfreq = "default",
  maxfreq = "default",
  mintime = "default",
  maxtime = "default",
  date,
  lat,
  lon,
  twilight = "sunlight",
  dawnstart = 0,
  dawnend = 5400,
  duskstart = 5400,
  duskend = 0,
  display = "plot",
  output = "percentage"
)
}
\arguments{
\item{aggregate_list}{The aggregated time-frequency dataframe produced by \code{\link{aggregate_df}}.}

\item{qvalues}{A numeric vector of sequential q-values for which to compute the diversity profiles, generally ranging between 0-5. \emph{e.g.:} \code{seq(0, 5, 0.1)} yields a vector between 0 and 5 with an increment value of 0.1. The smaller the increment value, the smoother the diversity profiles.}

\item{subset}{The scale for which the soundscape diversity is computed. Options are 'total', 'day',
'night', 'dawn', 'dusk' and 'tod' (time of day - for each unique time in the day).}

\item{minfreq}{A numeric value indicating the lower frequency limit for which to compute the soundscape diversity. If set to default, uses the lowest available frequency in the dataframe.}

\item{maxfreq}{A numeric value indicating the upper frequency limit for which to compute the soundscape diversity. If set to default, uses the highest available frequency in the dataframe.}

\item{mintime}{The lower time limit for which to compute the soundscape diversity, formatted as "HH:MM:SS". If set to default, uses the earliest time for which data exists in the dataframe.}

\item{maxtime}{The upper time limit for which to compute the soundscape diversity, formatted as "HH:MM:SS". If set to default, uses the latest time for which data exists in the dataframe.}

\item{date}{The first day of the recording period. Used for managing time-objects in R. Format as "YYYY-mm-dd".}

\item{lat}{The latitude of the site at which the sound files were collected, expressed in decimal degrees.}

\item{lon}{The longitude of the site at which the sound files were collected, expressed in decimal degrees.}

\item{twilight}{A character string of the twilight method to be used for sunrise and sunset as the boundary between day and night.
Options can be found in the \code{\link[photobiology]{day_night}} documentation.}

\item{dawnstart}{A numeric argument. If subset='dawn', used to determine the start of dawn. By default, dawn starts at sunrise. Expressed as the time in seconds before sunrise.}

\item{dawnend}{A numeric argument. If subset='dawn', used to determine the end of dawn. By default, dawn ends 1.5 hours after sunrise. Expressed as the time in seconds after sunrise.}

\item{duskstart}{A numeric argument. If subset='dusk', used to determine the start of dusk. By default, dusk starts 1.5 hours before sunset. Expressed as the time in seconds before sunset.}

\item{duskend}{A numeric argument. If subset='dusk', used to determine the end of dusk. By default, dusk ends at sunset. Expressed as the time in seconds after sunset.}

\item{display}{A character string. Indicates the output format, one of either "plot" or "table".}

\item{output}{A character string. Indicates the format in which the soundscape diversity is expressed. Options are "percentage" (the fraction between the observed soundscape diversity and the maximum possible soundscape diversity), or "raw" (the number of acoustically active OSUs in the soundscape). Defaults to "percentage".}
}
\value{
A soundscape diversity profile plot or table.
}
\description{
Gives the soundscape diversity as a continuous function of the order of diversity q. Soundscape diversity profiles are a useful visualization tool to characterize the abundance of OSUs in the system, and assess the evenness.
}
