% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundscapeR-package.R
\docType{package}
\name{soundscapeR-package}
\alias{soundscapeR}
\alias{soundscapeR-package}
\title{soundscapeR: Visualization And Quantification Of Soundscape Diversity Using Hill Numbers.}
\description{
The soundscapeR package constitutes a standardized analytical pipeline for the exploration, computation, visualization, and comparison of soundscape diveristy using Hill numbers. The package is designed to work with either continuous or regular-interval long-duration acoustic recordings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ThomasLuypaert/soundscapeR}
  \item Report bugs at \url{https://github.com/ThomasLuypaert/soundscapeR/issues}
}

}
\author{
\strong{Maintainer}: Thomas Luypaert \email{thomas.luypaert@nmbu.no} (\href{https://orcid.org/0000-0001-7491-7418}{ORCID})

Authors:
\itemize{
  \item Anderson Saldanha Bueno \email{buenoas@gmail.com} (\href{https://orcid.org/0000-0001-7416-6626}{ORCID})
  \item Tom Bradfer-Lawrence \email{tom.bradferlawrence@rspb.org.uk} (\href{https://orcid.org/0000-0001-6045-4360}{ORCID})
  \item Carlos Augusto Peres \email{carlos.a.peres@gmail.com} (\href{https://orcid.org/0000-0002-1588-8765}{ORCID})
  \item Torbjørn Haugaasen \email{torbjorn.haugaasen@nmbu.no} (\href{https://orcid.org/0000-0003-0901-5324}{ORCID})
}

}
\keyword{internal}
