% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_ss_index_calc.R
\name{ss_assess_files}
\alias{ss_assess_files}
\title{heck and clean the detected folders containing '.wav' files}
\usage{
ss_assess_files(file_locs, full_days = TRUE)
}
\arguments{
\item{file_locs}{The output of the `ss_find_files` function}

\item{full_days}{A boolean operator specifying whether the function should subset the files in each detected folder to full sampling days.
Defaults to TRUE.}
}
\value{
A named list with full-length paths to the '.wav' files contained in the parent directory or subdirectories after checking and cleaning.
}
\description{
This function takes the output list of the `ss_find_files` function and performs several checks on the
files. For instance, the function automatically detects the sampling regime of all files, and checks whether the time interval between
adjacent files in a folder deviates from the expected sampling regime (e.g. due to missing files). Moreover, the function
allows the user to subset the number of files per folder to contain only full sampling days (remove partially sampled days from the study).
}
