% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_ss_index_calc.R
\name{ss_index_calc}
\alias{ss_index_calc}
\title{Calculate the 'Acoustic Cover' (CVR) spectral acoustic index values for a folder of sound files}
\usage{
ss_index_calc(
  file_list,
  output_dir = NA,
  window = 256,
  theta = 3,
  threshold = 3
)
}
\arguments{
\item{file_list}{A list containing the full-length path to each '.wav' file in a folder. This list can be obtained using the
`ss_find_files` and `ss_assess_files` functions.}

\item{output_dir}{The full-length path to an output directory on your device. If not specified, will default to the location where
the raw sound files are saved.}

\item{window}{The window length used for the Fast-Fourier Transformation. Defaults to 256}

\item{theta}{For each neighbourhood (3 frames × 9 frequency bins) centred on any element/pixel in the spectrogram, calculate the average spectrogram value, ā. If ā is less than a user determined threshold, θ (theta), set the value of the central element/pixel equal to the minimum in the neighbourhood}

\item{threshold}{The cut-off dB-value above which sound is considered present in a frequency bin. Defaults to 3 dB.}
}
\value{
CVR-index output files in '.csv' form, stored in output_dir
}
\description{
For a given folder of sound files, this function is used to calculate the CVR-index for each 1-min file.
}
\examples{

# File preparation
fpath <- system.file("extdata", package="soundscapeR")
file_locs <- ss_find_files(parent_directory = fpath)
file_locs_clean <- ss_assess_files(file_locs = file_locs, full_days = FALSE)
# Index calculation
ss_index_calc(file_list = file_locs_clean[[1]], window = 256)

}
