% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_ss_index_calc.R
\name{ss_index_calc}
\alias{ss_index_calc}
\title{Calculate the 'Acoustic Cover' (CVR) spectral acoustic index values for a folder of sound files}
\usage{
ss_index_calc(
  file_list,
  output_dir = NA,
  window = 256,
  theta = 3,
  parallel = FALSE
)
}
\arguments{
\item{file_list}{A list containing the full-length path to each '.wav' file in a folder. This list can be obtained using the
`ss_find_files` and `ss_assess_files` functions.}

\item{output_dir}{The full-length path to an output directory on your device. If not specified, will default to the location where
the raw sound files are saved.}

\item{window}{The window length used for the Fast-Fourier Transformation. Defaults to 256}

\item{theta}{The cut-off dB-value above which sound is considered present in a frequency bin. Defaults to 3 dB.}

\item{parallel}{A boolean operator indicating whether parallel processing should be used to calculate the CVR-index files.
Defaults to FALSE.}
}
\value{
A list of CVR-values for each sound file
}
\description{
For a given folder of sound files, this function is used to calculate the CVR-index for each 1-min file.
}
