% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_function_names.R
\name{threshold_df}
\alias{threshold_df}
\title{threshold_df function}
\usage{
threshold_df(...)
}
\arguments{
\item{...}{no parameters}
}
\description{
Depreciated - Use 'ss_threshold' instead
}
